refactor: move parseConfigString to utils
